changes variables to unsigned long

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,7 +7,7 @@
 
 void print_number(int n)
 {
-	unsigned int digs, tens, absolute_n, control;
+	unsigned long int digs, tens, absolute_n, control;
 
 	absolute_n = n;
 	control = 1;
